// types/xmltv.i18n.ts
export type XmltvCreditImage = XmltvImage & {
  /**
   * Type of image
   */
  type?: 'person' | 'character';
};

/**
 * A representation of a person in an XMLTV programme object.
 */
export type XmltvPerson = {
  /**
   * The name of the person.
   */
  _value: string;

  /**
   * The role of the actor in the programme eg Bryan Cranston's role in Breaking Bad is "Walter White".
   */
  role?: string;

  /**
   * Whether the person is a guest star or a regular cast member.
   */
  guest?: boolean;

  /**
   * An image of the person.
   */
  image?: XmltvCreditImage[];

  /**
   * The URL of the person.
   */
  url?: XmltvUrl[];
};
// ...
export type XmltvCredits = {
  /**
   * The director(s) of the programme.
   */
  director?: XmltvPerson[];

  /**
   * The actor(s) in the programme.
   */
  actor?: XmltvPerson[];

  /**
   * The writer(s) of the programme.
   */
  writer?: XmltvPerson[];

  /**
   * The adapter(s) of the programme.
   */
  adapter?: XmltvPerson[];

  /**
   * The producer(s) of the programme.
   */
  producer?: XmltvPerson[];

  /**
   * The composer(s) of the programme.
   */
  composer?: XmltvPerson[];

  /**
   * The editor(s) of the programme.
   */
  editor?: XmltvPerson[];

  /**
   * The presenter(s) of the programme.
   */
  presenter?: XmltvPerson[];

  /**
   * The commentator(s) of the programme.
   */
  commentator?: XmltvPerson[];

  /**
   * The guest(s) of the programme.
   */
  guest?: XmltvPerson[];
};

/**
 * A way to define multiple strings with different langs for the same value.
 */
type XmltvStringWithLang = {
  /**
   * The value of the string
   */
  _value: string;

  /**
   * The lang code of the string eg. "en" for English, "es" for Spanish, etc.
   */
  lang?: string;
};

/**
 * The channel display name with an optional language attribute.
 */
export type XmltvDisplayName = XmltvStringWithLang;
// ...
export type XmltvIcon = {
  /**
   * URL of the icon
   */
  src: string;

  /**
   * Width of the icon
   */
  width?: number;

  /**
   * Height of the icon
   */
  height?: number;
};

/**
 * Object describing a URL, used in the programme and channel objects.
 */
export type XmltvUrl = {
  /**
   * Value of the URL
   */
  _value: string;

  /**
   * System of the URL, eg "imdb", "thetvdb", etc.
   */
  system?: string;
};
// ...
export type XmltvChannel = {
  /**
   * The channel id
   *
   * @example
   * ```typescript
   * {
   *   id: 'channel-1'
   * }
   * ```
   */
  id: string;

  /**
   * The channel display name
   *
   * @example
   * ```typescript
   * {
   *   displayName: [{
   *     _value: 'Channel 1',
   *     lang: 'en'
   *   },{
   *     _value: 'Kanal 1',
   *     lang: 'de'
   *   }]
   * }
   * ```
   */
  displayName: XmltvDisplayName[];

  /**
   * The channel icon
   *
   * @example
   * ```typescript
   * {
   *   icon: [{
   *     src: 'https://www.themoviedb.org/assets/1/v4/logos/32x32-blue-1f8b5c2fda197d0ee7d4f5b9fdca72a67ac3c9f7f8f8f8f8f8f8f8f8f8f8f8f8f.png',
   *     width: 32,
   *     height: 32
   *   }]
   * }
   * ```
   */
  icon?: XmltvIcon[];

  /**
   * The channel url
   *
   * @example
   * ```typescript
   * {
   *   url: [{_value: 'https://www.channel1.com'}]
   * ]
   * ```
   */
  url?: XmltvUrl[];
};

export type Xmltv = {
  channels?: XmltvChannel[];
  programmes?: XmltvProgramme[];
  date?: Date;
  sourceInfoName?: string;
  generatorInfoName?: string;
  sourceInfoUrl?: string;
  sourceDataUrl?: string;
  generatorInfoUrl?: string;
};